+++ saleor/graphql/checkout/types.py
@@ -1,11 +1,14 @@
 import graphene
 import graphene_django_optimizer as gql_optimizer
 from django.conf import settings
+from graphql_jwt.decorators import permission_required
 
 from ...checkout import models
 from ...checkout.utils import get_valid_shipping_methods_for_checkout
 from ...core.taxes import zero_taxed_money
 from ..core.connection import CountableDjangoObjectType
+from ..core.resolvers import resolve_meta, resolve_private_meta
+from ..core.types.meta import MetadataObjectType
 from ..core.types.money import TaxedMoney
 from ..giftcard.types import GiftCard
 from ..payment.enums import PaymentGatewayEnum
@@ -39,7 +42,7 @@ class CheckoutLine(CountableDjangoObjectType):
         return root.is_shipping_required()
 
 
-class Checkout(CountableDjangoObjectType):
+class Checkout(MetadataObjectType, CountableDjangoObjectType):
     available_shipping_methods = graphene.List(
         ShippingMethod,
         required=True,
@@ -159,3 +162,12 @@ class Checkout(CountableDjangoObjectType):
     @staticmethod
     def resolve_is_shipping_required(root: models.Checkout, _info):
         return root.is_shipping_required()
+
+    @staticmethod
+    @permission_required("order.manage_orders")
+    def resolve_private_meta(root, _info):
+        return resolve_private_meta(root, _info)
+
+    @staticmethod
+    def resolve_meta(root, _info):
+        return resolve_meta(root, _info)