+++ saleor/payment/utils.py
@@ -1,6 +1,5 @@
 import json
 import logging
-
 from decimal import Decimal
 from functools import wraps
 from typing import Dict, List
@@ -25,7 +24,6 @@ from . import (
     TransactionKind,
     get_payment_gateway,
 )
-
 from .interface import AddressData, GatewayResponse, PaymentData, TokenConfig
 from .models import Payment, Transaction
 
@@ -41,7 +39,7 @@ REQUIRED_GATEWAY_KEYS = {
     "currency",
 }
 ALLOWED_GATEWAY_KINDS = {choices[0] for choices in TransactionKind.CHOICES}
-GATEWAYS_META_LABEL = "gateways"
+GATEWAYS_META_NAMESPACE = "payment-gateways"
 
 
 def list_enabled_gateways() -> List[str]:
@@ -515,22 +513,24 @@ def gateway_refund(payment, amount: Decimal = None) -> Transaction:
 
 def fetch_customer_id(user, gateway):
     """Retrieves users customer_id stored for desired gateway"""
-    key = prepare_label_name(gateway)
-    gateway_config = user.get_private_meta(label=GATEWAYS_META_LABEL).get(key, {})
+    key = prepare_namespace_name(gateway)
+    gateway_config = user.get_private_meta(
+        namespace=GATEWAYS_META_NAMESPACE, client=key
+    )
     return gateway_config.get("customer_id", None)
 
 
 def store_customer_id(user, gateway, customer_id):
     """Stores customer_id in users private meta for desired gateway"""
     user.store_private_meta(
-        label=GATEWAYS_META_LABEL,
-        key=prepare_label_name(gateway),
-        value={"customer_id": customer_id},
+        namespace=GATEWAYS_META_NAMESPACE,
+        client=prepare_namespace_name(gateway),
+        item={"customer_id": customer_id},
     )
     user.save(update_fields=["private_meta"])
 
 
-def prepare_label_name(s):
+def prepare_namespace_name(s):
     return s.strip().upper()