+++ saleor/static/dashboard-next/components/ProductList/ProductList.tsx
@@ -12,10 +12,10 @@ import TableRow from "@material-ui/core/TableRow";
 import React from "react";
 
 import TableCellAvatar from "@saleor/components/TableCellAvatar";
+import i18n from "@saleor/i18n";
+import { maybe, renderCollection } from "@saleor/misc";
+import { ListActions, ListProps } from "@saleor/types";
 import { CategoryDetails_category_products_edges_node } from "../../categories/types/CategoryDetails";
-import i18n from "../../i18n";
-import { maybe, renderCollection } from "../../misc";
-import { ListActions, ListProps } from "../../types";
 import Checkbox from "../Checkbox";
 import Money from "../Money";
 import Skeleton from "../Skeleton";
@@ -71,6 +71,7 @@ interface ProductListProps
 export const ProductList = withStyles(styles, { name: "ProductList" })(
   ({
     classes,
+    settings,
     disabled,
     isChecked,
     pageInfo,
@@ -81,6 +82,7 @@ export const ProductList = withStyles(styles, { name: "ProductList" })(
     toolbar,
     onNextPage,
     onPreviousPage,
+    onUpdateListSettings,
     onRowClick
   }: ProductListProps) => (
     <Table>
@@ -109,8 +111,10 @@ export const ProductList = withStyles(styles, { name: "ProductList" })(
         <TableRow>
           <TablePagination
             colSpan={6}
+            settings={settings}
             hasNextPage={pageInfo && !disabled ? pageInfo.hasNextPage : false}
             onNextPage={onNextPage}
+            onUpdateListSettings={onUpdateListSettings}
             hasPreviousPage={
               pageInfo && !disabled ? pageInfo.hasPreviousPage : false
             }