+++ saleor/static/dashboard-next/pages/components/PageList/PageList.tsx
@@ -17,9 +17,9 @@ import Skeleton from "@saleor/components/Skeleton";
 import StatusLabel from "@saleor/components/StatusLabel";
 import TableHead from "@saleor/components/TableHead";
 import TablePagination from "@saleor/components/TablePagination";
-import i18n from "../../../i18n";
-import { maybe, renderCollection } from "../../../misc";
-import { ListActions, ListProps } from "../../../types";
+import i18n from "@saleor/i18n";
+import { maybe, renderCollection } from "@saleor/misc";
+import { ListActions, ListProps } from "@saleor/types";
 import { PageList_pages_edges_node } from "../../types/PageList";
 
 export interface PageListProps extends ListProps, ListActions {
@@ -47,11 +47,13 @@ const styles = (theme: Theme) =>
 const PageList = withStyles(styles, { name: "PageList" })(
   ({
     classes,
+    settings,
     pages,
     disabled,
     onNextPage,
     pageInfo,
     onRowClick,
+    onUpdateListSettings,
     onPreviousPage,
     isChecked,
     selected,
@@ -82,8 +84,10 @@ const PageList = withStyles(styles, { name: "PageList" })(
           <TableRow>
             <TablePagination
               colSpan={4}
+              settings={settings}
               hasNextPage={pageInfo && !disabled ? pageInfo.hasNextPage : false}
               onNextPage={onNextPage}
+              onUpdateListSettings={onUpdateListSettings}
               hasPreviousPage={
                 pageInfo && !disabled ? pageInfo.hasPreviousPage : false
               }