+++ saleor/static/dashboard-next/shipping/components/ShippingZonesList/ShippingZonesList.tsx
@@ -20,10 +20,10 @@ import Checkbox from "@saleor/components/Checkbox";
 import Skeleton from "@saleor/components/Skeleton";
 import TableHead from "@saleor/components/TableHead";
 import TablePagination from "@saleor/components/TablePagination";
-import i18n from "../../../i18n";
-import { maybe, renderCollection } from "../../../misc";
-import { ICONBUTTON_SIZE } from "../../../theme";
-import { ListActions, ListProps } from "../../../types";
+import i18n from "@saleor/i18n";
+import { maybe, renderCollection } from "@saleor/misc";
+import { ICONBUTTON_SIZE } from "@saleor/theme";
+import { ListActions, ListProps } from "@saleor/types";
 import { ShippingZoneFragment } from "../../types/ShippingZoneFragment";
 
 export interface ShippingZonesListProps extends ListProps, ListActions {
@@ -54,10 +54,12 @@ const ShippingZonesList = withStyles(styles, { name: "ShippingZonesList" })(
   ({
     classes,
     disabled,
+    settings,
     onAdd,
     onNextPage,
     onPreviousPage,
     onRemove,
+    onUpdateListSettings,
     onRowClick,
     pageInfo,
     shippingZones,
@@ -98,8 +100,10 @@ const ShippingZonesList = withStyles(styles, { name: "ShippingZonesList" })(
           <TableRow>
             <TablePagination
               colSpan={4}
+              settings={settings}
               hasNextPage={pageInfo && !disabled ? pageInfo.hasNextPage : false}
               onNextPage={onNextPage}
+              onUpdateListSettings={onUpdateListSettings}
               hasPreviousPage={
                 pageInfo && !disabled ? pageInfo.hasPreviousPage : false
               }