+++ CHANGELOG.md
@@ -22,6 +22,10 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix dashboard menu styles - #4491 by @benekex2
 - Do not allow random ids to appear in snapshots - #4495 by @dominik-zeglen
 - Order is no longer created when the payment was unsuccessful in the API - #4500 by @NyanKiyoshi
+- Fix navigation rwd - #4511 by @benekex2
+- Create general abstraction for object metadata - #4447 by @salwator
+- Contrast improvements - #4508 by @benekex2
+- Allow selecting the number of rows displayed in dashboard's list views - #4414 by @benekex2
 
 ## 2.8.0